Extract to Extend class
