+++ apps/bring-app/plugins/bring-app/src/Extend/Extend.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BringApp\Extend;
+
+use BringApp\Extend\JWTAuth\JWTAuth;
+
+class Extend {
+	/**
+	 * Initialize plugin extensions
+	 */
+	public static function init() {
+		JWTAuth::init();
+	}
+}
