+++ apps/bring-app/plugins/bring-app/src/Extend/JWTAuth/index.php
@@ -1,0 +1,4 @@
+<?php
+
+// No direct access
+defined("ABSPATH") or die("Hey, do not do this 😱");
